Refactor descriptors
Removed unnecessary abstraction level to increase the codebase easiness and modifiability
