+++ generator/src/main/java/org/sudu/protogen/descriptors/EnumOrMessage.java
@@ -1,4 +1,4 @@
-package org.sudu.protogen.protobuf;
+package org.sudu.protogen.descriptors;
 
 import org.apache.commons.lang3.Validate;
 import org.jetbrains.annotations.NotNull;
