+++ generator/src/main/java/org/sudu/protogen/descriptors/GenerationRequest.java
@@ -1,4 +1,4 @@
-package org.sudu.protogen.protobuf;
+package org.sudu.protogen.descriptors;
 
 import java.util.List;
