+++ generator/src/main/java/org/sudu/protogen/descriptors/Method.java
@@ -1,43 +1,88 @@
-package org.sudu.protogen.protoc.adaptor;
+package org.sudu.protogen.descriptors;
 
 import com.google.protobuf.Descriptors;
-import org.sudu.protogen.protobuf.RepeatedContainer;
+import org.apache.commons.lang3.Validate;
 import org.sudu.protogen.protoc.Options;
 
 import java.util.Objects;
 import java.util.Optional;
 
-public class Method extends org.sudu.protogen.protobuf.Method {
+public class Method {
 
-    Descriptors.MethodDescriptor descriptor;
+    private final Descriptors.MethodDescriptor descriptor;
 
     public Method(Descriptors.MethodDescriptor descriptor) {
         this.descriptor = descriptor;
     }
 
-    @Override
     public String getName() {
         return descriptor.getName();
     }
 
-    @Override
     public Message getInputType() {
         return new Message(descriptor.getInputType());
     }
 
-    @Override
     public Message getOutputType() {
         return new Message(descriptor.getOutputType());
     }
 
-    @Override
     public boolean isInputStreaming() {
         return descriptor.isClientStreaming();
     }
 
-    @Override
     public boolean isOutputStreaming() {
         return descriptor.isServerStreaming();
+    }
+
+
+    public final boolean isNullable() {
+        return getNullableOption().orElse(false);
+    }
+
+    public final String generatedName() {
+        return getNameOption().orElseGet(this::getName);
+    }
+
+    public final boolean doGenerate() {
+        if (isInputStreaming()) return false;
+        return getGenerateOption()
+                .orElse(true);
+    }
+
+    public final boolean doUnfoldRequest() {
+        return getUnfoldRequestOption()
+                .orElse(!getInputType().isDomain());
+    }
+
+    public final boolean doUnfoldResponse() {
+        return !getOutputType().isDomain() && getOutputType().getFields().size() == 1;
+    }
+
+    public final Field unfoldedResponseField() {
+        Validate.validState(doUnfoldResponse());
+        return getOutputType().getFields().get(0);
+    }
+
+    protected Optional<Boolean> getGenerateOption() {
+        return Options.wrapExtension(descriptor.getOptions(), protogen.Options.genMethod);
+    }
+
+    protected Optional<Boolean> getUnfoldRequestOption() {
+        return Options.wrapExtension(descriptor.getOptions(), protogen.Options.unfoldRequest);
+    }
+
+    protected Optional<Boolean> getNullableOption() {
+        return Options.wrapExtension(descriptor.getOptions(), protogen.Options.nullable);
+    }
+
+    public Optional<RepeatedContainer> getStreamToContainer() {
+        return Options.wrapExtension(descriptor.getOptions(), protogen.Options.streamToContainer)
+                .map(RepeatedContainer::fromGrpc);
+    }
+
+    protected Optional<String> getNameOption() {
+        return Options.wrapExtension(descriptor.getOptions(), protogen.Options.methodName);
     }
 
     @Override
@@ -51,31 +96,5 @@ public class Method extends org.sudu.protogen.protobuf.Method {
     @Override
     public int hashCode() {
         return Objects.hash(descriptor);
-    }
-
-    @Override
-    protected Optional<Boolean> getGenerateOption() {
-        return Options.wrapExtension(descriptor.getOptions(), protogen.Options.genMethod);
-    }
-
-    @Override
-    protected Optional<Boolean> getUnfoldRequestOption() {
-        return Options.wrapExtension(descriptor.getOptions(), protogen.Options.unfoldRequest);
-    }
-
-    @Override
-    protected Optional<Boolean> getNullableOption() {
-        return Options.wrapExtension(descriptor.getOptions(), protogen.Options.nullable);
-    }
-
-    @Override
-    public Optional<RepeatedContainer> getStreamToContainer() {
-        return Options.wrapExtension(descriptor.getOptions(), protogen.Options.streamToContainer)
-                .map(RepeatedContainer::fromGrpc);
-    }
-
-    @Override
-    protected Optional<String> getNameOption() {
-        return Options.wrapExtension(descriptor.getOptions(), protogen.Options.methodName);
     }
 }
