+++ generator/src/main/java/org/sudu/protogen/generator/GenerationContext.java
@@ -2,11 +2,11 @@ package org.sudu.protogen.generator;
 
 import com.squareup.javapoet.TypeSpec;
 import org.sudu.protogen.config.Configuration;
+import org.sudu.protogen.descriptors.EnumOrMessage;
+import org.sudu.protogen.descriptors.File;
+import org.sudu.protogen.descriptors.Service;
 import org.sudu.protogen.generator.field.processors.FieldTypeProcessor;
 import org.sudu.protogen.generator.type.processors.TypeProcessor;
-import org.sudu.protogen.protobuf.EnumOrMessage;
-import org.sudu.protogen.protobuf.File;
-import org.sudu.protogen.protobuf.Service;
 
 import java.util.List;
 import java.util.Map;
