+++ generator/src/main/java/org/sudu/protogen/generator/TypeTable.java
@@ -5,8 +5,8 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.sudu.protogen.ProtogenException;
 import org.sudu.protogen.config.Configuration;
-import org.sudu.protogen.protobuf.EnumOrMessage;
-import org.sudu.protogen.protobuf.File;
+import org.sudu.protogen.descriptors.EnumOrMessage;
+import org.sudu.protogen.descriptors.File;
 import org.sudu.protogen.utils.Name;
 
 import java.util.Arrays;
