+++ generator/src/main/java/org/sudu/protogen/generator/client/MethodGeneratorBase.java
@@ -2,13 +2,13 @@ package org.sudu.protogen.generator.client;
 
 import com.squareup.javapoet.*;
 import org.sudu.protogen.ProtogenException;
+import org.sudu.protogen.descriptors.Field;
+import org.sudu.protogen.descriptors.Method;
 import org.sudu.protogen.generator.GenerationContext;
 import org.sudu.protogen.generator.field.FieldGenerator;
 import org.sudu.protogen.generator.type.IteratorType;
 import org.sudu.protogen.generator.type.RepeatedType;
 import org.sudu.protogen.generator.type.TypeModel;
-import org.sudu.protogen.protobuf.Field;
-import org.sudu.protogen.protobuf.Method;
 
 import java.util.List;
