+++ generator/src/main/java/org/sudu/protogen/generator/field/processors/DomainFieldTypeProcessor.java
@@ -1,9 +1,9 @@
 package org.sudu.protogen.generator.field.processors;
 
 import org.jetbrains.annotations.NotNull;
+import org.sudu.protogen.descriptors.Field;
 import org.sudu.protogen.generator.GenerationContext;
 import org.sudu.protogen.generator.type.TypeModel;
-import org.sudu.protogen.protobuf.Field;
 
 public class DomainFieldTypeProcessor extends FieldTypeProcessor.Chain {
