+++ generator/src/main/java/org/sudu/protogen/generator/field/processors/ListFieldTypeProcessor.java
@@ -1,10 +1,10 @@
 package org.sudu.protogen.generator.field.processors;
 
 import org.jetbrains.annotations.NotNull;
+import org.sudu.protogen.descriptors.Field;
 import org.sudu.protogen.generator.GenerationContext;
 import org.sudu.protogen.generator.type.RepeatedType;
 import org.sudu.protogen.generator.type.TypeModel;
-import org.sudu.protogen.protobuf.Field;
 
 class ListFieldTypeProcessor extends FieldTypeProcessor.Chain {
