+++ generator/src/main/java/org/sudu/protogen/generator/field/processors/MapFieldTypeProcessor.java
@@ -1,11 +1,11 @@
 package org.sudu.protogen.generator.field.processors;
 
 import org.jetbrains.annotations.NotNull;
+import org.sudu.protogen.descriptors.Field;
+import org.sudu.protogen.descriptors.Message;
 import org.sudu.protogen.generator.GenerationContext;
 import org.sudu.protogen.generator.type.MapType;
 import org.sudu.protogen.generator.type.TypeModel;
-import org.sudu.protogen.protobuf.Field;
-import org.sudu.protogen.protobuf.Message;
 
 class MapFieldTypeProcessor extends FieldTypeProcessor.Chain {
