+++ generator/src/main/java/org/sudu/protogen/generator/field/processors/UnfoldedFieldTypeProcessor.java
@@ -1,12 +1,12 @@
 package org.sudu.protogen.generator.field.processors;
 
 import org.jetbrains.annotations.NotNull;
+import org.sudu.protogen.descriptors.Field;
 import org.sudu.protogen.generator.GenerationContext;
 import org.sudu.protogen.generator.type.PrimitiveTypeModel;
 import org.sudu.protogen.generator.type.RepeatedType;
 import org.sudu.protogen.generator.type.TypeModel;
 import org.sudu.protogen.generator.type.UnfoldedType;
-import org.sudu.protogen.protobuf.Field;
 import org.sudu.protogen.utils.Poem;
 
 class UnfoldedFieldTypeProcessor extends FieldTypeProcessor.Chain {
