+++ generator/src/main/java/org/sudu/protogen/generator/type/processors/TypeProcessor.java
@@ -3,9 +3,9 @@ package org.sudu.protogen.generator.type.processors;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.sudu.protogen.ProtogenException;
+import org.sudu.protogen.descriptors.EnumOrMessage;
 import org.sudu.protogen.generator.GenerationContext;
 import org.sudu.protogen.generator.type.TypeModel;
-import org.sudu.protogen.protobuf.EnumOrMessage;
 
 import java.util.List;
